ISO I/O library : imbecile crime of the century!!!
The p1 compiler does not permit any arithmetic on values of type FilePos nor does it permit conversion to another type. As ridiculous as this may seem, it does not violate the ISO standard which defines type FilePos as an array and does not require it to support arithmetic nor conversion. We therefore have no choice but to open the file, read it byte by byte to the end while incrementing a counter to obtain the filesize as a value of a useful type. This is terribly inefficiant, especially on larger files. The penalty for using such a badly designed standard.